Resolve "Improved model 'physical_non_linearity_with_saturation'"

Closes #596

See merge request esa/pyxel!655

+++ pyxel/detectors/detector.py
@@ -338,7 +338,7 @@ class Detector:
             raise ValueError("No readout defined.")
 
     @property
-    def pipeline_count(self) -> float:
+    def pipeline_count(self) -> int:
         """TBW."""
         if self._readout_properties is not None:
             return self._readout_properties.pipeline_count
